requirements.txt
aiogram==3.2.0
ccxt==3.0.75
fastapi==0.107.0
uvicorn==0.22.0
sqlalchemy==2.0.19
alembic==1.11.1
pandas==2.2.2
numpy==1.26.2
python-dotenv==1.0.0
apscheduler==3.10.1
httpx==0.24.1
pydantic==2.5.1
aiofiles==23.2.1
aiohttp==3.9.5
certifi==2025.8.3
magic-filter==1.0.12
